pyproject.toml
[tool.poetry]
name = "teletweet"
version = "0.1.0"
description = ""
authors = ["marinhocesar <marinhocaiocesar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.14.0"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.4"
gunicorn = "^23.0.0"
flask = {extras = ["async"], version = "^3.0.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
